Add `test_viz`
