+++ tests/__init__.py
@@ -2,8 +2,8 @@
 Test package.
 """
 
-from . import (test__cache, test_dbms, test_dirs, test_geom, test_ops, test_settings, test_store,
-               test_text)
+from . import test__cache, test_dbms, test_dirs, test_geom, test_ops, test_settings, test_store, \
+    test_text, test_viz
 
 __all__ = [
     'test__cache',
@@ -14,4 +14,5 @@ __all__ = [
     'test_settings',
     'test_store',
     'test_text',
+    'test_viz',
 ]
