+++ tests/test_ops/test_manipulation.py
@@ -4,8 +4,6 @@ Tests the :mod:`~pyhelpers.ops.manipulation` submodule.
 
 import typing
 
-import matplotlib.cm
-import matplotlib.pyplot as plt
 import pytest
 
 from pyhelpers._cache import _lazy_check_dependencies, example_dataframe
@@ -246,20 +244,6 @@ def test_flatten_columns():
             ['A_B_C', 'A_B_D'])
     assert flatten_columns([('A', 'C'), ('B', 'C')], ignore_linear_level=True) == ['A', 'B']
     assert flatten_columns([('A',), ('B', 'C')], ignore_linear_level=True) == ['A', 'B']
-
-
-def test_cmap_discretisation():
-    cm_accent = cmap_discretisation(matplotlib.colormaps['Accent'], n_colours=5)
-    assert cm_accent.name == 'Accent_5'
-
-
-def test_colour_bar_index():
-    plt.figure(figsize=(2, 6))
-
-    cbar = colour_bar_index(cmap=matplotlib.colormaps['Accent'], n_colours=5, labels=list('abcde'))
-    cbar.ax.tick_params(labelsize=14)
-
-    plt.close(fig='all')
 
 
 if __name__ == '__main__':
